+++ src/Resources/skeleton/tree/doctrine/Repository.tpl.php
@@ -1,0 +1,13 @@
+<?= "<?php\n"; ?>
+
+namespace <?= $namespace; ?>;
+
+use Gedmo\Tree\Entity\Repository\NestedTreeRepository;
+
+/**
+ *
+ */
+class <?= $class_name; ?> extends NestedTreeRepository
+{
+
+}